Added Account Controller
Throwing System.AggregateException
TODO: Fix the aforementioned exception

+++ FitDeck.Web/FitDeck.Models/Account/ApplicationUserCreate.cs
@@ -23,7 +23,7 @@ namespace FitDeck.Models.Account
         public string LastName { get; set; }
 
         [Required(ErrorMessage = "Height is required")]
-        public float Height { get; set; }
+        public int Height { get; set; }
 
         [Required(ErrorMessage = "Weight is required")]
         public int Weight { get; set; }
